Refactoring + bug fix
