+++ MapsInMyFolder.Common/Database.cs
@@ -298,7 +298,7 @@ namespace MapsInMyFolder.Commun
             }
         }
 
-        //CONNEXION A LA BASE DE DONNEES
+
         public static SQLiteConnection DB_Connection()
         {
             // Create a new database connection:
@@ -328,7 +328,6 @@ namespace MapsInMyFolder.Commun
         public static SQLiteConnection DB_OpenConnection(string datasource)
         {
             SQLiteConnection sqlite_conn = new SQLiteConnection("Data Source=" + datasource + "; Version = 3; New = True; Compress = True; ");
-            // Open the connection:
             try
             {
                 sqlite_conn.Open();
@@ -492,7 +491,8 @@ namespace MapsInMyFolder.Commun
             if (IsUpdateSuccessful)
             {
                 string ExistingDatabasePath = Path.Combine(Settings.working_folder, Settings.database_pathname);
-                if (File.Exists(ExistingDatabasePath)) {
+                if (File.Exists(ExistingDatabasePath))
+                {
                     string backupFolderPath = Path.Combine(Settings.working_folder, "databaseBackup");
                     Directory.CreateDirectory(backupFolderPath);
                     string dateFormatee = DateTime.Now.ToString("yyyy-MM-dd-HH-mm-ss");
