+++ MapsInMyFolder.Common/packages.config
@@ -13,11 +13,10 @@
   <package id="Mapbox.VectorTile" version="1.0.4-alpha2" targetFramework="net472" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="7.0.0" targetFramework="net472" />
   <package id="ModernWpfUis" version="1.4.0" targetFramework="net472" />
-  <package id="NetVips" version="2.3.1" targetFramework="net472" />
+  <package id="NetVips" version="2.4.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net472" />
   <package id="ProjNET" version="2.0.0" targetFramework="net472" />
   <package id="SkiaSharp" version="2.88.6" targetFramework="net472" />
-  <!-- <package id="SkiaSharp.NativeAssets.macOS" version="2.88.6" targetFramework="net472" />-->
   <package id="SkiaSharp.NativeAssets.Win32" version="2.88.6" targetFramework="net472" />
   <package id="Stub.System.Data.SQLite.Core.NetFramework" version="1.0.118.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
