+++ MapsInMyFolder.sln
@@ -63,8 +63,8 @@ Global
 		{579455BC-EE2B-4B6E-BF0C-CB3FEAC2D69D}.Release|x64.Build.0 = Release|x64
 		{579455BC-EE2B-4B6E-BF0C-CB3FEAC2D69D}.Release|x86.ActiveCfg = Release|x64
 		{579455BC-EE2B-4B6E-BF0C-CB3FEAC2D69D}.Release|x86.Build.0 = Release|x64
-		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|Any CPU.Build.0 = Debug|x64
 		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|x64.ActiveCfg = Debug|x64
 		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|x64.Build.0 = Debug|x64
 		{ADC22697-3B59-47FA-B8CD-FC75AB6012E3}.Debug|x86.ActiveCfg = Debug|x64
