+++ MapsInMyFolder/LayersPannel.cs
@@ -8,15 +8,12 @@ using System.Data.SQLite;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 using System.Text.RegularExpressions;
-using System.Threading;
 using System.Threading.Tasks;
 using System.Web;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Threading;
 
 namespace MapsInMyFolder
 {
@@ -112,67 +109,8 @@ namespace MapsInMyFolder
             {
                 try
                 {
-                    SQLiteDataReader sqlite_datareaderCopy = sqlite_datareader;
-                    string GetStringFromOrdinal(string name)
-                    {
-                        return Database.GetStringFromOrdinal(sqlite_datareaderCopy, name);
-                    }
-                    int? GetIntFromOrdinal(string name)
-                    {
-                        return Database.GetIntFromOrdinal(sqlite_datareaderCopy, name);
-                    }
-
-                    int? DB_Layer_ID = GetIntFromOrdinal("ID");
-                    string DB_Layer_NAME = GetStringFromOrdinal("NAME").RemoveNewLineChar();
-                    bool DB_Layer_FAVORITE = Convert.ToBoolean(GetIntFromOrdinal("FAVORITE"));
-                    string DB_Layer_DESCRIPTION = GetStringFromOrdinal("DESCRIPTION");
-                    string DB_Layer_CATEGORY = GetStringFromOrdinal("CATEGORY").RemoveNewLineChar();
-                    string DB_Layer_COUNTRY = GetStringFromOrdinal("COUNTRY").RemoveNewLineChar();
-                    string DB_Layer_IDENTIFIER = GetStringFromOrdinal("IDENTIFIER").RemoveNewLineChar();
-                    string DB_Layer_TILE_URL = GetStringFromOrdinal("TILE_URL").RemoveNewLineChar();
-                    int? DB_Layer_MIN_ZOOM = GetIntFromOrdinal("MIN_ZOOM");
-                    int? DB_Layer_MAX_ZOOM = GetIntFromOrdinal("MAX_ZOOM");
-                    string DB_Layer_FORMAT = GetStringFromOrdinal("FORMAT");
-                    string DB_Layer_SITE = GetStringFromOrdinal("SITE").RemoveNewLineChar();
-                    string DB_Layer_SITE_URL = GetStringFromOrdinal("SITE_URL").RemoveNewLineChar();
-                    string DB_Layer_STYLE = GetStringFromOrdinal("STYLE");
-                    int? DB_Layer_TILE_SIZE = GetIntFromOrdinal("TILE_SIZE");
-                    string DB_Layer_SCRIPT = GetStringFromOrdinal("SCRIPT");
-                    string DB_Layer_VISIBILITY = GetStringFromOrdinal("VISIBILITY");
-                    string DB_Layer_SPECIALSOPTIONS = GetStringFromOrdinal("SPECIALSOPTIONS");
-                    string DB_Layer_RECTANGLES = GetStringFromOrdinal("RECTANGLES");
-                    int DB_Layer_VERSION = GetIntFromOrdinal("VERSION") ?? 0;
-                    bool DB_Layer_HAS_SCALE = Convert.ToBoolean(GetIntFromOrdinal("HAS_SCALE"));
-
-                    bool doCreateSpecialsOptionsClass = true;
-                    Layers.LayersSpecialsOptions DeserializeSpecialsOptions = null;
-                    try
-                    {
-                        if (!string.IsNullOrEmpty(DB_Layer_SPECIALSOPTIONS))
-                        {
-                            DeserializeSpecialsOptions = System.Text.Json.JsonSerializer.Deserialize<Layers.LayersSpecialsOptions>(DB_Layer_SPECIALSOPTIONS);
-                            doCreateSpecialsOptionsClass = false;
-                        }
-                    }
-                    catch (Exception)
-                    {
-                        Debug.WriteLine("Invalide JSON inside layer :" + DB_Layer_ID + " named : " + DB_Layer_NAME);
-                    }
-                    finally
-                    {
-                        if (doCreateSpecialsOptionsClass)
-                        {
-                            DeserializeSpecialsOptions = new Layers.LayersSpecialsOptions();
-                        }
-                    }
-
-                    if (!string.IsNullOrEmpty(DB_Layer_SCRIPT))
-                    {
-                        DB_Layer_SCRIPT = Collectif.HTMLEntities(DB_Layer_SCRIPT, true);
-                    }
-
-                    Layers calque = new Layers((int)DB_Layer_ID, DB_Layer_FAVORITE, DB_Layer_NAME, DB_Layer_DESCRIPTION, DB_Layer_CATEGORY, DB_Layer_COUNTRY, DB_Layer_IDENTIFIER, DB_Layer_TILE_URL, DB_Layer_SITE, DB_Layer_SITE_URL, DB_Layer_MIN_ZOOM, DB_Layer_MAX_ZOOM, DB_Layer_FORMAT, DB_Layer_STYLE, DB_Layer_TILE_SIZE, DB_Layer_SCRIPT, DB_Layer_VISIBILITY, DeserializeSpecialsOptions, DB_Layer_RECTANGLES, DB_Layer_VERSION, DB_Layer_HAS_SCALE);
-                    if (DB_Layer_FAVORITE && Settings.layerpanel_favorite_at_top)
+                    var calque = Layers.GetLayerFromSQLiteDataReader(sqlite_datareader);
+                    if (calque.IsFavorite && Settings.layerpanel_favorite_at_top)
                     {
                         layersFavorite.Add(calque);
                     }
@@ -203,7 +141,7 @@ namespace MapsInMyFolder
 
             List<Layers> legacyLayers = LayerReadInDatabase(OriginalLayersGetQuery);
             List<Layers> editedLayers = LayerReadInDatabase(EditedLayersGetQuery);
-            LayersMergeLegacyWithEdited(legacyLayers, editedLayers);
+            Layers.LayersMergeLegacyWithEdited(legacyLayers, editedLayers);
             string baseHTML = LayersCreateHTML();
 
             if (Settings.show_layer_devtool)
@@ -225,98 +163,6 @@ namespace MapsInMyFolder
 
         }
 
-
-        static void LayersMergeLegacyWithEdited(List<Layers> legacyLayers, List<Layers> editedLayers)
-        {
-            Layers.Clear();
-            Dictionary<int, Layers> editedLayersDictionnary = editedLayers.ToDictionary(l => l.Id, l => l);
-            foreach (Layers legacyLayer in legacyLayers)
-            {
-                int legacyLayerVersion = legacyLayer.Version;
-                bool legacyLayerHasReplacement = editedLayersDictionnary.TryGetValue(legacyLayer.Id, out Layers replacementLayer);
-
-                BindingFlags fieldsBindingFlags = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.Static;
-                Layers legacyLayerWithReplacements = legacyLayer;
-                if (legacyLayerHasReplacement)
-                {
-                    foreach (FieldInfo field in typeof(Layers).GetFields(fieldsBindingFlags))
-                    {
-                        object replacementValue = field.GetValue(replacementLayer);
-                        if (replacementValue is null)
-                        {
-                            continue;
-                        }
-                        Type replacementValueType = replacementValue.GetType();
-
-                        if (replacementValueType == typeof(string))
-                        {
-                            if (replacementValue is string replacementValueTypeToString)
-                            {
-                                field.SetValue(legacyLayerWithReplacements, replacementValueTypeToString);
-                            }
-                        }
-                        else
-                        {
-                            field.SetValue(legacyLayerWithReplacements, replacementValue);
-                        }
-                    }
-
-
-                    if (legacyLayer.Version > replacementLayer.Version)
-                    {
-                        legacyLayerWithReplacements.DoShowWarningLegacyVersionNewerThanEdited = true;
-                    }
-                }
-
-                if (legacyLayerWithReplacements?.Visibility?.Trim() == "DELETED")
-                {
-                    continue;
-                }
-
-                if (string.IsNullOrEmpty(legacyLayerWithReplacements.Script))
-                {
-                    legacyLayerWithReplacements.Script = Settings.tileloader_default_script;
-                }
-                if (string.IsNullOrEmpty(legacyLayerWithReplacements.Visibility))
-                {
-                    legacyLayerWithReplacements.Visibility = "Visible";
-                }
-                if (legacyLayerWithReplacements.Category == "/")
-                {
-                    legacyLayerWithReplacements.Category = "";
-                }
-                legacyLayerWithReplacements.Country = legacyLayerWithReplacements.Country?.Replace("*", "World");
-                List<string> listOfAllFormatsAcceptedWithTransparency = new List<string> { "png" };
-                if (!string.IsNullOrWhiteSpace(legacyLayerWithReplacements.TilesFormat) && listOfAllFormatsAcceptedWithTransparency.Contains(legacyLayerWithReplacements.TilesFormat))
-                {
-                    legacyLayerWithReplacements.TilesFormatHasTransparency = true;
-                }
-
-                //make sure there is no null values inside the layer
-                foreach (FieldInfo field in typeof(Layers).GetFields(fieldsBindingFlags))
-                {
-                    object actualValue = field.GetValue(legacyLayerWithReplacements);
-                    if (actualValue is null)
-                    {
-                        if (field.FieldType == typeof(string))
-                        {
-                            field.SetValue(legacyLayerWithReplacements, string.Empty);
-                        }
-                        else if (field.FieldType == typeof(int))
-                        {
-                            field.SetValue(legacyLayerWithReplacements, 0);
-                        }
-                        else
-                        {
-                            field.SetValue(legacyLayerWithReplacements, null);
-                        }
-                    }
-                }
-
-                Layers.Add(Convert.ToInt32(legacyLayerWithReplacements.Id), legacyLayerWithReplacements);
-
-            }
-        }
 
         static string LayersCreateHTML()
         {
@@ -865,128 +711,8 @@ namespace MapsInMyFolder
                 Debug.WriteLine("Error getPreview exception :" + ex.Message);
             }
             return String.Empty;
-
         }
     }
 
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("Performance", "CA1822:Marquer les membres comme étant static", Justification = "Used by CEFSHARP")]
-    public class LayerCEFSharpLink
-    {
-        public void ClearCache(string listOfId = "0")
-        {
-            long DirectorySize = 0;
-            string[] splittedListOfId = listOfId.Split(',', StringSplitOptions.RemoveEmptyEntries);
-            foreach (string str in splittedListOfId)
-            {
-                int id_int = int.Parse(str.Trim());
-                DirectorySize += MainPage.ClearCache(id_int);
-                Debug.WriteLine("Clear_cache layer " + id_int);
-            }
-            if (DirectorySize >= 0)
-            {
-                string memoryFreed = Collectif.FormatBytes(DirectorySize);
-                string cacheCleanedMessage = "";
 
-                if (splittedListOfId.Length == 1)
-                {
-                    cacheCleanedMessage = Languages.GetWithArguments("layerMessageCacheCleared", Layers.GetLayerById(int.Parse(splittedListOfId[0])).Name, memoryFreed);
-                }
-                else
-                {
-                    cacheCleanedMessage = Languages.GetWithArguments("layerMessageCachesCleared", Layers.GetLayerById(int.Parse(splittedListOfId[0])).Name, memoryFreed);
-                }
-                Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-                {
-                    Message.NoReturnBoxAsync(cacheCleanedMessage, Languages.Current["dialogTitleOperationSuccess"]);
-                }, null);
-            }
-        }
-
-
-        public void LayerFavorite(double id = 0, bool isAdding = true)
-        {
-            int id_int = Convert.ToInt32(id);
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-            {
-                MainPage.DBLayerFavorite(id_int, isAdding);
-            }, null);
-        }
-
-        public void LayerVisibility(double id = 0, bool isVisible = true)
-        {
-            //Debug.WriteLine($"Layer_visibility : id={id} & isVisible={isVisible}");
-            int id_int = Convert.ToInt32(id);
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-            {
-                MainPage.DBLayerVisibility(id_int, (isVisible ? "Visible" : "Hidden"));
-            }, null);
-        }
-
-
-        public void LayerMakeEdits(double id = 0, double prefilid = -1)
-        {
-            int id_int = Convert.ToInt32(id);
-            int prefilid_int = Convert.ToInt32(prefilid);
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-            {
-                MainWindow.Instance.FrameLoad_CustomOrEditLayers(id_int, prefilid_int);
-            }, null);
-        }
-
-        public void LayerSetAsCurrent(double id = 0)
-        {
-            int id_int = Convert.ToInt32(id);
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-            {
-                //Debug.WriteLine("Layer_set_current " + id);
-                MainWindow.Instance.MainPage.SetCurrentLayer(id_int);
-            }, null);
-        }
-        public void LayerShowWarningLegacyVersionNewerThanEdited(double id = 0)
-        {
-            int id_int = Convert.ToInt32(id);
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-            {
-                MainPage.ShowLayerWarning(id_int);
-            }, null);
-        }
-
-        public void LayerRequestSearchUpdate()
-        {
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Send, (SendOrPostCallback)delegate
-            {
-                MainWindow.Instance.MainPage.SearchLayerStart(true);
-            }, null);
-        }
-
-        public string LayerRequestGetSearchString()
-        {
-            return Application.Current.Dispatcher.Invoke(() => MainWindow.Instance.MainPage.SearchGetText(), DispatcherPriority.Send);
-        }
-
-        public void LayerRequestRefreshPanel()
-        {
-            Application.Current.Dispatcher.BeginInvoke(DispatcherPriority.Normal, (SendOrPostCallback)delegate
-            {
-                MainWindow.Instance.MainPage.ReloadPage();
-                MainWindow.Instance.MainPage.SearchLayerStart();
-            }, null);
-        }
-
-        public string LayerGetTilePreviewUrlFromId(double id = 0)
-        {
-            if (!Settings.layerpanel_livepreview)
-            {
-                return "";
-            }
-            async Task<string> LayerGetTilePreviewUrlFromId_Task(double id)
-            {
-                int id_int = Convert.ToInt32(id);
-                DispatcherOperation op = Application.Current.Dispatcher.BeginInvoke(new Func<string>(() => MainWindow.Instance.MainPage.LayerTilePreview_ReturnUrl(id_int)));
-                await op;
-                return op.Result.ToString();
-            }
-            return LayerGetTilePreviewUrlFromId_Task(id).Result;
-        }
-    }
 }
